Move conversion functions into types files
